+++ example/lib/home_page.dart
@@ -23,8 +23,8 @@ class HomePage extends StatelessWidget {
                   builder: (context) {
                     return CodePage(
                       title: 'Complete Form',
-                      child: CompleteForm(),
                       sourceFilePath: 'lib/sources/complete_form.dart',
+                      child: CompleteForm(),
                     );
                   },
                 ),
@@ -41,8 +41,8 @@ class HomePage extends StatelessWidget {
                   builder: (context) {
                     return CodePage(
                       title: 'Signup Form',
-                      child: SignupForm(),
                       sourceFilePath: 'lib/sources/signup_form.dart',
+                      child: SignupForm(),
                     );
                   },
                 ),