+++ example/lib/sources/complete_form.dart
@@ -408,33 +408,6 @@ class CompleteFormState extends State<CompleteForm> {
                     border: Border.all(color: Colors.green),
                     onChanged: _onChanged,
                   ),
-                  FormBuilderImagePicker(
-                    name: 'photos',
-                    decoration: const InputDecoration(labelText: 'Pick Photos'),
-                    maxImages: 1,
-                  ),
-                  const SizedBox(height: 15),
-                  FormBuilderPhoneField(
-                    name: 'phone_number',
-                    decoration: const InputDecoration(
-                      border: OutlineInputBorder(),
-                      labelText: 'Phone Number',
-                      hintText: 'Hint',
-                    ),
-                    onChanged: _onChanged,
-                    priorityListByIsoCode: ['KE'],
-                    validator: FormBuilderValidators.compose([
-                      FormBuilderValidators.numeric(context),
-                      FormBuilderValidators.required(context),
-                    ]),
-                  ),
-                  const SizedBox(height: 15),
-                  FormBuilderFilePicker(
-                    name: 'files',
-                    previewImages: false,
-                    decoration: InputDecoration(labelText: 'Files filed'),
-                  ),
-                  const SizedBox(height: 15),
                 ],
               ),
             ),
@@ -443,10 +416,6 @@ class CompleteFormState extends State<CompleteForm> {
                 Expanded(
                   child: MaterialButton(
                     color: Theme.of(context).accentColor,
-                    child: const Text(
-                      'Submit',
-                      style: TextStyle(color: Colors.white),
-                    ),
                     onPressed: () {
                       if (_formKey.currentState.saveAndValidate()) {
                         print(_formKey.currentState.value);
@@ -455,19 +424,23 @@ class CompleteFormState extends State<CompleteForm> {
                         print('validation failed');
                       }
                     },
+                    child: const Text(
+                      'Submit',
+                      style: TextStyle(color: Colors.white),
+                    ),
                   ),
                 ),
                 const SizedBox(width: 20),
                 Expanded(
-                  child: OutlineButton(
-                    color: Theme.of(context).accentColor,
+                  child: OutlinedButton(
+                    onPressed: () {
+                      _formKey.currentState.reset();
+                    },
+                    // color: Theme.of(context).accentColor,
                     child: Text(
                       'Reset',
                       style: TextStyle(color: Theme.of(context).accentColor),
                     ),
-                    onPressed: () {
-                      _formKey.currentState.reset();
-                    },
                   ),
                 ),
               ],