+++ lib/l10n/messages_all.dart
@@ -15,33 +15,43 @@ import 'package:intl/intl.dart';
 import 'package:intl/message_lookup_by_library.dart';
 import 'package:intl/src/intl_helpers.dart';
 
+import 'messages_de.dart' as messages_de;
 import 'messages_en.dart' as messages_en;
 import 'messages_es.dart' as messages_es;
 import 'messages_fr.dart' as messages_fr;
+import 'messages_it.dart' as messages_it;
 import 'messages_ja.dart' as messages_ja;
 import 'messages_messages.dart' as messages_messages;
+import 'messages_pl.dart' as messages_pl;
 import 'messages_pt.dart' as messages_pt;
 import 'messages_sk.dart' as messages_sk;
 
 typedef Future<dynamic> LibraryLoader();
 Map<String, LibraryLoader> _deferredLibraries = {
+  'de': () => new Future.value(null),
   'en': () => new Future.value(null),
   'es': () => new Future.value(null),
   'fr': () => new Future.value(null),
+  'it': () => new Future.value(null),
   'ja': () => new Future.value(null),
   'messages': () => new Future.value(null),
   'pt': () => new Future.value(null),
   'sk': () => new Future.value(null),
+  'pl': () => new Future.value(null),
 };
 
 MessageLookupByLibrary _findExact(String localeName) {
   switch (localeName) {
+    case 'de':
+      return messages_de.messages;
     case 'en':
       return messages_en.messages;
     case 'es':
       return messages_es.messages;
     case 'fr':
       return messages_fr.messages;
+    case 'it':
+      return messages_it.messages;
     case 'ja':
       return messages_ja.messages;
     case 'messages':
@@ -50,6 +60,8 @@ MessageLookupByLibrary _findExact(String localeName) {
       return messages_pt.messages;
     case 'sk':
       return messages_sk.messages;
+    case 'pl':
+      return messages_pl.messages;
     default:
       return null;
   }