+++ lib/l10n/messages_sk.dart
@@ -19,15 +19,15 @@ typedef String MessageIfAbsent(String messageStr, List<dynamic> args);
 class MessageLookup extends MessageLookupByLibrary {
   String get localeName => 'sk';
 
-  static m0(value) => "Hodnota tohto poľa musí byť ${value}.";
+  static m4(value) => "Hodnota tohto poľa musí byť ${value}.";
 
-  static m1(max) => "Hodnota musí byť menšia alebo rovná ako ${max}.";
+  static m0(max) => "Hodnota musí byť menšia alebo rovná ako ${max}.";
 
-  static m2(maxLength) => "Hodnota musí mať dĺžku najviac ${maxLength} znakov.";
+  static m1(maxLength) => "Hodnota musí mať dĺžku najviac ${maxLength} znakov.";
 
-  static m3(min) => "Hodnota musí byť väčšia alebo rovná ako ${min}.";
+  static m2(min) => "Hodnota musí byť väčšia alebo rovná ako ${min}.";
 
-  static m4(minLength) => "Hodnota musí mať dĺžku aspoň ${minLength} znakov.";
+  static m3(minLength) => "Hodnota musí mať dĺžku aspoň ${minLength} znakov.";
 
   static m5(value) => "Hodnota tohto poľa nesmie byť ${value}.";
 
@@ -39,17 +39,17 @@ class MessageLookup extends MessageLookupByLibrary {
             "Toto pole vyžaduje platný dátum."),
         "emailErrorText": MessageLookupByLibrary.simpleMessage(
             "Toto pole vyžaduje platnú emailovú adresu."),
-        "equalErrorText": m0,
+        "equalErrorText": m4,
         "integerErrorText": MessageLookupByLibrary.simpleMessage(
             "Hodnota musí byť celé číslo."),
         "ipErrorText": MessageLookupByLibrary.simpleMessage(
             "Toto pole vyžaduje platnú IP adresu."),
         "matchErrorText": MessageLookupByLibrary.simpleMessage(
             "Hodnota nevyhovuje očakávanému tvaru."),
-        "maxErrorText": m1,
-        "maxLengthErrorText": m2,
-        "minErrorText": m3,
-        "minLengthErrorText": m4,
+        "maxErrorText": m0,
+        "maxLengthErrorText": m1,
+        "minErrorText": m2,
+        "minLengthErrorText": m3,
         "notEqualErrorText": m5,
         "numericErrorText":
             MessageLookupByLibrary.simpleMessage("Hodnota musí byť číslo."),